test: added options to test pixijs by adding webgl mocking fix: rename Constants.tsx to .ts
