+++ tsconfig.json
@@ -16,7 +16,6 @@
   "include": [
     "src/**/*.ts",
     "src/**/*.d.ts",
-    "src/**/*.tsx",
     "src/**/*.vue",
     "components.d.ts"
   ],
